variant: Workaround unimplemented 'mangling nontype_argument_pack'
In gcc-4.7 20111231, the previous definition of 'apply_funcs_walker'
will cause an 'unimplemented: mangling nontype_argument_pack' error. The
structure is now rewritten so that its member is no longer a template,
to avoid the error.

Also, added an explicit destructor for generic_unique_ptr to avoid the
inlining warnings.