event_loop: Added simple callback and ensure double-cancel is harmless.
The event registrar should be overloaded to take 2 versions of
functions. One takes less parameters, targeting for the common cases,
and one takes all necessary parameters, targeting for the generic case.

Also, enforcing double-cancel on the same handle must do no-op, so that
we don't need to keep track of the unique_event's. Thus the libev
implementation changed the type of event_handle from variant to an
integral type (plus other changes to adopt for the simple callbacks).